refactor: drag end drop to new component
