+++ src/components/ConstructorGame/DragEndDrop/DragEndDrop.module.css
@@ -1,0 +1,27 @@
+.word__wrapper {
+  display: flex;
+  list-style-type: none;
+  justify-content: center;
+
+  margin: 1rem 0;
+  width: 100%;
+}
+
+.word__char {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  margin: 0 1rem;
+  height: 50px;
+  width: 60px;
+
+  background: #733999;
+  border-radius: 10px;
+
+  color: #ffffff;
+  font-size: 2.4rem;
+  line-height: 2.8rem;
+  text-decoration: none;
+  cursor: pointer;
+}
